Hopefully the final round of data class refactoring - started with alliance and guild

+++ commands/alliance/delete.js
@@ -4,6 +4,7 @@ const ROOT = '../..';
 
 // Load our classes
 const Alliance = require(`${ROOT}/modules/alliance/Alliance`);
+const Guild    = require(`${ROOT}/modules/alliance/Guild`);
 
 // Load singletons
 const client = require(`${ROOT}/modules/Client`); // eslint-disable-line no-unused-vars
@@ -32,21 +33,27 @@ const run = async (message, args, level) => {
     }
     
     const value = args.join(' ').replace(/^"(.+)"$/g, "$1").replace(/^'(.+)'$/g, "$1");
-    const alliances = await Alliance.getByNameOrAlias({alliance_name: value, alliance_alias: value});
     
+    const alliances = await Alliance.get({nameOrShortName: {name: value, shortName: value}});
     if (alliances.length == 0) {
         message.channel.send(`Cannot find an alliance with that name or alias`);
         return;
     }
-    
     const alliance = alliances[0];
+    
+    // Check if there are any guilds in the alliance
+    const guilds = await Guild.get({allianceId: alliance.id});
+    if (guilds.length > 0) {
+        message.channel.send(`Cowardly refusing to delete alliance while there are still clan discords in it`);
+        return;
+    }
     
     try {
         await alliance.delete();
         message.channel.send(`Alliance deleted`);
     
     } catch (error) {
-        const label = `${alliance.alliance_name} [${alliance.alliance_alias}]`;
+        const label = `${alliance.name} [${alliance.shortName}]`;
         client.replyWithErrorAndDM(`Deletion of alliance failed: ${label}`, message, error);
     }
 };
